move pages folder out of component
